fix: 하위 필드 아이템 컴포넌트 수정
